Change filters from usecase to entities
