+++ app/usecase/host/get_localhost_usecase.go
@@ -1,8 +1,6 @@
 package host
 
 import (
-	"errors"
-
 	"github.com/PaoGRodrigues/tfi-backend/app/domain/host"
 )
 
@@ -16,6 +14,15 @@ func NewGetLocalhostsUseCase(huc host.HostUseCase) *GetLocalhostsUseCase {
 	}
 }
 
+func (l *GetLocalhostsUseCase) GetLocalHosts() ([]host.Host, error) {
+	current, err := l.checkHosts()
+	if err != nil {
+		return []host.Host{}, err
+	}
+	localHosts := host.GetLocalHosts(current)
+	return localHosts, nil
+}
+
 func (l *GetLocalhostsUseCase) checkHosts() ([]host.Host, error) {
 	current := l.searcher.GetHosts()
 	if len(current) == 0 {
@@ -26,52 +33,4 @@ func (l *GetLocalhostsUseCase) checkHosts() ([]host.Host, error) {
 		current = hosts
 	}
 	return current, nil
-}
-
-func (l *GetLocalhostsUseCase) GetLocalHosts() ([]host.Host, error) {
-	current, err := l.checkHosts()
-	if err != nil {
-		return []host.Host{}, err
-	}
-	localHosts := []host.Host{}
-	if len(current) != 0 {
-		for _, host := range current {
-			if host.PrivateHost {
-				localHosts = append(localHosts, host)
-			}
-		}
-	}
-	return localHosts, nil
-}
-
-func (l *GetLocalhostsUseCase) GetRemoteHosts() ([]host.Host, error) {
-	current, err := l.checkHosts()
-	if err != nil {
-		return []host.Host{}, err
-	}
-	remoteHosts := []host.Host{}
-	if len(current) != 0 {
-		for _, host := range current {
-			if !host.PrivateHost {
-				remoteHosts = append(remoteHosts, host)
-			}
-		}
-	}
-	return remoteHosts, nil
-}
-
-func (l *GetLocalhostsUseCase) GetHost(attr string) (host.Host, error) {
-	current, err := l.checkHosts()
-	if err != nil {
-		return host.Host{}, err
-	}
-	var ip string
-	for _, host := range current {
-		if host.IP == attr || host.Name == attr {
-			return host, nil
-		}
-		ip = host.IP
-	}
-
-	return host.Host{}, errors.New("There's no host with this IP " + ip)
 }
