+++ app/usecase/host/get_localhost_usecase_test.go
@@ -74,132 +74,3 @@ func TestGetLocalHostAndGetAllHostsInSearcherReturnError(t *testing.T) {
 		t.Fail()
 	}
 }
-
-func TestGetRemoteHostAndCallGetAllHostsInSearcherReturnRemoteHostsSuccessfully(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return(expected)
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	got, err := filter.GetRemoteHosts()
-	if err != nil {
-		t.Fail()
-	}
-
-	assert.Equal(t, []host.Host{remote}, got)
-}
-
-func TestGetRemoteHostCallingGetHostFromRepoInSearcherReturnRemoteHosts(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return([]host.Host{})
-	mockSearcher.EXPECT().GetAllHosts().Return(expected, nil)
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	got, err := filter.GetRemoteHosts()
-	if err != nil {
-		t.Fail()
-	}
-
-	assert.Equal(t, []host.Host{remote}, got)
-}
-
-func TestGetRemoteHostAndGetAllHostsInSearcherReturnAnError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return([]host.Host{})
-	mockSearcher.EXPECT().GetAllHosts().Return(nil, fmt.Errorf("Testing Error"))
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	_, err := filter.GetRemoteHosts()
-
-	if err == nil {
-		t.Fail()
-	}
-}
-
-func TestGetHostByIPReturnCorrectHost(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return(expected)
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	got, err := filter.GetHost(local.IP)
-
-	if err != nil {
-		t.Fail()
-	}
-
-	assert.Equal(t, local, got)
-}
-
-func TestGetHostByIPGetAllHostsReturnError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return(nil)
-	mockSearcher.EXPECT().GetAllHosts().Return(nil, fmt.Errorf("Error Test"))
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	_, err := filter.GetHost(local.IP)
-
-	if err == nil {
-		t.Fail()
-	}
-}
-
-func TestGetHostByIPWithAnUnexistingIPReturnError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return(expected)
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	_, err := filter.GetHost("10.10.10.10")
-
-	if err == nil {
-		t.Fail()
-	}
-}
-
-func TestGetHostByURLReturnCorrectHost(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return(expected)
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	got, err := filter.GetHost(remote.Name)
-
-	if err != nil {
-		t.Fail()
-	}
-
-	assert.Equal(t, remote, got)
-}
-
-func TestGetHostByAnUnexistingURLReturnError(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockSearcher := mocks.NewMockHostUseCase(ctrl)
-	mockSearcher.EXPECT().GetHosts().Return(expected)
-
-	filter := usecase.NewGetLocalhostsUseCase(mockSearcher)
-	_, err := filter.GetHost("test.test.com")
-
-	if err == nil {
-		t.Fail()
-	}
-}
